build with warnings on; assert rejections in the test suite
Add -Wall -Wextra -Wswitch-enum and fix the two resulting warnings:
strlen returns size_t and was compared against an int length.

-Wswitch-enum rather than plain -Wswitch because codegen's switch has a
default: case, which suppresses -Wswitch entirely; a node kind added
without a codegen case would otherwise go unreported.

test.sh had no shebang (#/bin/bash, missing the !), no way to assert
that bad input is rejected, and left temp files behind. Add assert_fail,
clean up per assertion, and report a count.

Fix the stale @file in parse.c's header comment.

+++ parse.c
@@ -1,5 +1,5 @@
 /**
- * @file ycc.c
+ * @file parse.c
  * @author Yusuke Ohashi(mail@yusuke.cloud)
  * @brief 
  * @version 0.1
@@ -48,7 +48,7 @@ void error_at(char *loc, char *fmt, ...) {
 
 bool consume(char *op) {
     if (token->kind != TK_RESERVED || 
-    strlen(op) != token->len ||
+    (int)strlen(op) != token->len ||
     /* if they match each other, it's 0(false). Anything that is not 0 is true in C. */
     memcmp(token->str, op, token->len)) {
         return false;
@@ -60,7 +60,7 @@ bool consume(char *op) {
 
 void expect(char *op) {
     if (token->kind != TK_RESERVED ||
-    strlen(op) != token->len ||
+    (int)strlen(op) != token->len ||
     memcmp(token->str, op, token->len)) {
         error_at(token->str, "This is not '%s'", op);
     }
